Modify the structure
