+++ Dockerfile
@@ -1,4 +1,4 @@
-FROM python:3.8
+FROM python:3.9.12
 
 # Setting up the working directory
 WORKDIR /app
