+++ edge_server.py
@@ -1,10 +1,7 @@
 import pickle
 import numpy as np
 import tensorflow as tf
-import pynvml
-import psutil
 import socket
-import time as tm
 from model import AlexNet
 import os
 
@@ -36,15 +33,6 @@ def preprocess_image(image, label):
     image /= 255.0  # Normalize to [0, 1]
     label = tf.one_hot(label, depth=10)
     return image, label
-# def make_dataset(single_sample=False):
-#     # Load the CIFAR-10 dataset
-#     (train_images, train_labels), _ = tf.keras.datasets.cifar10.load_data()
-#     train_dataset = tf.data.Dataset.from_tensor_slices((train_images, train_labels))
-#     train_dataset = train_dataset.map(preprocess_image)
-#     if not single_sample:
-#         train_dataset = train_dataset.batch(128).shuffle(1000)
-#     return train_dataset
-# /app/data/cifar-10-batches-py/
 def make_dataset(single_sample=False, data_dir="cifar-10-batches-py/"):
     train_images, train_labels = load_cifar10_data(data_dir)
     train_dataset = tf.data.Dataset.from_tensor_slices((train_images, train_labels))
@@ -76,60 +64,6 @@ def calculate_averages(data):
             averages[key].append(average_at_index)
 
     return averages
-def get_time(alexNet, data_set):
-    iterator = iter(data_set)
-    image, label = next(iterator)
-    label = tf.expand_dims(label, axis=0)  # 添加批次维度
-    # Add Batch Dimension，because the model expects four-dimensional inputs (batch_size, height, width, channels)
-    image = tf.expand_dims(image, axis=0)
-    training_times = {
-        "forward": [],
-        "backward": [],
-        "update": []
-    }
-    with tf.GradientTape(persistent=True) as tape:
-        x = image
-        intermediates = []
-        for layer in alexNet.model.layers:
-            start_time = tm.time()
-            x = layer(x)
-            intermediates.append(x)
-            training_times['forward'].append(tm.time() - start_time)
-        loss = tf.reduce_mean(tf.losses.categorical_crossentropy(label, x, from_logits=False))
-
-    for idx, layer in enumerate(alexNet.model.layers):
-        if layer.trainable_variables:  # Only compute for layers with parameters
-            start_time = tm.time()
-            grads = tape.gradient(loss, layer.trainable_variables)
-            training_times["backward"].append(tm.time() - start_time)
-            # 更新参数
-            start_time = tm.time()
-            tf.keras.optimizers.Adam().apply_gradients(zip(grads, layer.trainable_variables))
-            training_times["update"].append(tm.time() - start_time)
-        else:
-            training_times["backward"].append(0)
-            training_times["update"].append(0)
-
-    return training_times
-# Get the performance of the current device
-# def get_device_performance():
-#     # Get the number of gpu of the current device
-#     # pynvml.nvmlInit()
-#     # gpu_count = pynvml.nvmlDeviceGetCount()
-#     # print("GPU count:", gpu_count)
-#     # pynvml.nvmlShutdown()
-#
-#     # Get the number of cpu of the current device
-#     # cpu_count = psutil.cpu_count(logical=True)
-#     cpu_count = 8
-#     print("logical CPUs count:", cpu_count)
-#     # Get the RAM size of the current device
-#     # ram_info = psutil.virtual_memory()
-#     # ram = ram_info.total / (1024 ** 3)
-#     ram = 4
-#     print("RAM: ", ram, "GB")
-
-
 def send_edge_server_info(data, host, port):
     with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
         s.connect((host, port))
@@ -253,7 +187,7 @@ if __name__ == "__main__":
     print("Dataset loaded.")
     train_times_list = []
     for i in range(50):
-        train_times_list.append(get_time(model, single_sample_data_set))
+        train_times_list.append(model.get_time(single_sample_data_set))
     train_times = calculate_averages(train_times_list)
     print(train_times)
     # score = get_device_performance()
